Reorganized packages and file locations
