+++ src/main/java/com/chess/api/game/Colour.java
@@ -1,0 +1,6 @@
+package com.chess.api.game;
+
+public enum Colour {
+    WHITE,
+    BLACK;
+}
