+++ src/main/java/com/chess/api/game/condition/Conditional.java
@@ -1,0 +1,10 @@
+package com.chess.api.game.condition;
+
+import com.chess.api.game.Board;
+import com.chess.api.game.movement.Action;
+
+public interface Conditional {
+
+    boolean isExpected(Board board, Action action);
+
+}
