+++ src/main/java/com/chess/api/game/movement/MovementType.java
@@ -1,0 +1,6 @@
+package com.chess.api.game.movement;
+
+public enum MovementType {
+    JUMP,
+    ADVANCE
+}
