+++ src/main/java/com/chess/api/game/piece/PieceType.java
@@ -1,0 +1,20 @@
+package com.chess.api.game.piece;
+
+public enum PieceType {
+    PAWN(""),
+    ROOK("R"),
+    KNIGHT("N"),
+    BISHOP("B"),
+    QUEEN("Q"),
+    KING("K");
+
+    private final String code;
+
+    PieceType(String code) {
+        this.code = code;
+    }
+
+    public String getCode() {
+        return code;
+    }
+}
