+++ src/main/java/com/chess/api/game/reference/Direction.java
@@ -1,0 +1,9 @@
+package com.chess.api.game.reference;
+
+public enum Direction {
+    AT,
+    FRONT,
+    BACK,
+    LEFT,
+    RIGHT
+}
