+++ src/main/java/com/chess/api/game/reference/Location.java
@@ -1,0 +1,10 @@
+package com.chess.api.game.reference;
+
+public enum Location {
+    LAST_MOVED,
+    START,
+    DESTINATION,
+    VECTOR,
+    PATH_TO_DESTINATION,
+    PATH_TO_VECTOR,
+}
