--- src/main/java/com/chess/api/model/Action.java
@@ -1,5 +1,0 @@
-package com.chess.api.model;
-
-public record Action(Colour colour, Vector2D start, Vector2D end) {
-
-}
