--- src/main/java/com/chess/api/model/Colour.java
@@ -1,6 +1,0 @@
-package com.chess.api.model;
-
-public enum Colour {
-    WHITE,
-    BLACK;
-}
