--- src/main/java/com/chess/api/model/movement/MovementType.java
@@ -1,6 +1,0 @@
-package com.chess.api.model.movement;
-
-public enum MovementType {
-    JUMP,
-    ADVANCE
-}
