--- src/main/java/com/chess/api/model/movement/condition/Conditional.java
@@ -1,10 +1,0 @@
-package com.chess.api.model.movement.condition;
-
-import com.chess.api.model.Action;
-import com.chess.api.model.Board;
-
-public interface Conditional {
-
-    boolean isExpected(Board board, Action action);
-
-}
