--- src/main/java/com/chess/api/model/movement/condition/Direction.java
@@ -1,9 +1,0 @@
-package com.chess.api.model.movement.condition;
-
-public enum Direction {
-    AT,
-    FRONT,
-    BACK,
-    LEFT,
-    RIGHT
-}
