--- src/main/java/com/chess/api/model/movement/condition/Location.java
@@ -1,10 +1,0 @@
-package com.chess.api.model.movement.condition;
-
-public enum Location {
-    LAST_MOVED,
-    START,
-    DESTINATION,
-    VECTOR,
-    PATH_TO_DESTINATION,
-    PATH_TO_VECTOR,
-}
