--- src/main/java/com/chess/api/model/piece/PieceType.java
@@ -1,20 +1,0 @@
-package com.chess.api.model.piece;
-
-public enum PieceType {
-    PAWN(""),
-    ROOK("R"),
-    KNIGHT("N"),
-    BISHOP("B"),
-    QUEEN("Q"),
-    KING("K");
-
-    private final String code;
-
-    PieceType(String code) {
-        this.code = code;
-    }
-
-    public String getCode() {
-        return code;
-    }
-}
