+++ src/test/java/com/chess/api/game/GameTest.java
@@ -1,0 +1,5 @@
+package com.chess.api.game;
+
+class GameTest {
+
+}
