--- src/test/java/com/chess/api/model/GameTest.java
@@ -1,5 +1,0 @@
-package com.chess.api.model;
-
-class GameTest {
-
-}
